FinalProjectHarvardHappinessHub
First copy of final project. Started with a freely available theme.
Added design elements (adjusting fonts, locations, title, and
background, which was a photo taken on-campus and edited in Picasa
software). Added links to necessary pages; will add more in the future.
Embedded a video on the top of the page.