Website as for now
I first seeked a title that works with Harvard (our popularity) and
Happiness. I noticed they both begin with “H” so I searched online for
works beginning with “H” - wow there are so many of them. FInally
decided “Harvard Happiness Hub”. Cool. The next part was to start with
a freely available theme by choosing a tumblr template and then using
the skills I learned at class to tweak it to what I want. I will work
on it more!
 I added design elements: adjusted the fonts, locations, title, and
more.
I then walked for 3 hours (believe it or not!) around Harvard Campus to
grab the moments of sun and seek a photo to be the background. I edited
it in Picasa software.
I  also added links to necessary pages and will add more in the future.
Finally I embedded a video on the top of the page.