Final week edits
Final week updates, on http://harvardhappiness.shalhavit.com/
My goals were (quoting from last week): 
"To make changes in writing and understand the “making it live” process
and edit all the things (or at least a bunch) of my “To Do” list that I
accumulated this week: changing URLs, Changing text, and overall
understanding GoDaddy a bit."
So: Yes! I figured it out and now I feel more confident with the
web publishing process and making edits.
"My ‘DONE’ Next week: Well, I hope to fix some technicalities in the
video exporting process as they came out with pixelized parts. As far
as web design goes, I would like to try and add the HashTag option
(Bakhtiar suggested wisely that I do so through twitter - and perhaps
in future to upload to twitter immediately) and/or add a Comment option
underneath the videos. Either this by next week or create another page
just for comments. 
It might be the best thing to at least create a document /jpg of the
vision so even if it’s not all ready by next week I can continue
working on it and have a clear goal in mind.
I would also like to get more user feedback this week from those who
submitted videos for my project and those who watched it. 
I want to present this interactive learning platform format to my boss
at EDx and therefore I want to be sure that my website is presentable
and that the vision is clear enough even if the site is yet imperfect."

Ok so what I got done this week is:
1) Home Page -> Added subscribe button. (Completely functional, even
email ID is stored in Database). Also added in all pages in footer.
2) About Page -> Updated with 3 new videos that I re-edited b-roll to
give a clearer message.
3) Explore Page: 
     a) Added search bar - though it's currently in "Beta". You can
enter some search term text but it will redirect to "under
construction" page. But at least it shows what I want it to be in the
future. It won't be too much work to make this change live soon.
     b) Similarly: "submit video" also redirects to "under construction
page" (I tried learning from the TUMBLR, it worked but I did something
and now it's still not working so this is the first thing that I should
fix after class). 
     c) Featured videos and other videos for all four weeks updated
-YES! Also the data is stored in database. I will need to create an
entry in database if I want new/change/remove video.
4) Related Links - No change made. In the future I will add to this and
hopefully get insight from students through the "contact" page. Or
ideally, I will find a way to make it self monitored like Wikipedia.
5) Contact - My users can fill in a form and click submit. Completely
functional. I receive the "contact" emails with details of the person's
name, email ID and what message he/she wrote and it is directed to my
personal email.

I think the website overall looks OK. Of course the visuals can always
be better and I ALWAYS seem to want to fix and change them which takes
forever so I will leave it looking like this for now. I've made it as
clean as possible and most important: it's functional!
Also I tested it on a few friends and they liked looking around,
watching the videos, and interacting with the website. I got very good
feedback.
It is still in beta mode and there is more work to do. But looking back
I learned so much. I never thought I could do anything like this
before. Code used to be total Chinese for me and it's unbelievable that
I ACTUALLY GET things now. I learned that web design is very much about
learning from others, communicating well, accumulating knowledge and
experience and slowly things become clear and achievable. 

Thank you Bakhtiar and Alec.
There are two old Jewish proverbs "A wise person is that who learns
from every human" and "make practicing become your habit" (Chapters of
the Fathers). This is so relevant in web design and web development.
Practice and practice, fail and learn. And learn! I am grateful and I
thank all my inspiring colleagues, you're priceless!